requirements.txt
asgiref==3.7.1
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.0.10
django-allauth==0.48.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-spectacular==0.21.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
